opa-ff: add wildcard whitelisting (bsc#1174464)
This script contains dynamic no-op content in comments that changes with
every release. We would either need to whitelisting each of these no-op
changes or we would need to adapt our digesting mechanism to skip '#'
comment lines. Since this instance if currently the only case of this
I'm taking the wildcard whitelisting route for the moment.